+++ version4.go
@@ -4,7 +4,11 @@
 
 package uuid
 
-import "io"
+import (
+	"encoding/binary"
+	"io"
+	chacha8RandV2 "math/rand/v2"
+)
 
 // New creates a new random UUID or panics.  New is equivalent to
 // the expression
@@ -37,6 +41,10 @@ func NewString() string {
 //	equivalent to the odds of creating a few tens of trillions of UUIDs in a
 //	year and having one duplicate.
 func NewRandom() (UUID, error) {
+	if _, ok := rander.(defaultRandReader); ok {
+		return fastRandV4DefaultZeroAlloc(), nil
+	}
+
 	if !poolEnabled {
 		return NewRandomFromReader(rander)
 	}
@@ -73,4 +81,37 @@ func newRandomFromPool() (UUID, error) {
 	uuid[6] = (uuid[6] & 0x0f) | 0x40 // Version 4
 	uuid[8] = (uuid[8] & 0x3f) | 0x80 // Variant is 10
 	return uuid, nil
+}
+
+// Singleton with no state that implements
+// io.Reader and uses the default rand/v2
+// package to read bytes.
+type defaultRandReader struct{}
+
+// Read fills the provided byte slice `b` with random data using a
+// cryptographically secure random number generator.
+func (d defaultRandReader) Read(b []byte) (n int, err error) {
+	var num uint64
+	numByteIndex := 0
+
+	for i := range b {
+		if numByteIndex == 0 {
+			num = chacha8RandV2.Uint64()
+		}
+		b[i] = byte(num >> (numByteIndex * 8))
+		numByteIndex = (numByteIndex + 1) % 8
+	}
+
+	return len(b), nil
+}
+
+func fastRandV4DefaultZeroAlloc() UUID {
+	var uuid UUID
+	hi, low := chacha8RandV2.Uint64(), chacha8RandV2.Uint64()
+	binary.LittleEndian.PutUint64(uuid[0:8], hi)
+	binary.LittleEndian.PutUint64(uuid[8:16], low)
+
+	uuid[6] = (uuid[6] & 0x0f) | 0x40 // Version 4
+	uuid[8] = (uuid[8] & 0x3f) | 0x80 // Variant is 10
+	return uuid
 }
